nka/requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.14
filelock==3.7.1
gcloud==0.18.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
jwcrypto==1.3.1
oauth2client==4.1.3
platformdirs==2.5.2
protobuf==4.21.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==3.3.3
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
urllib3==1.26.10
virtualenv==20.15.1
virtualenvwrapper-win==1.2.7
wrapt==1.14.1
zipp==3.8.1